--- apps/skeleton-app/src/_gqlTypes/GET_ATTRIBUTES_BY_LIB.ts
@@ -1,20 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-/* tslint:disable */
-/* eslint-disable */
-// @generated
-// This file was automatically generated and should not be edited.
-
-import {AttributeFormat} from './globalTypes';
-
-// ====================================================
-// GraphQL query operation: GET_ATTRIBUTES_BY_LIB
-// ====================================================
-
-
-export interface GET_ATTRIBUTES_BY_LIB_attributes_list_StandardAttribute_embedded_fields {
-    id: string;
-    format: AttributeFormat | null;
-    label: SystemTranslation | null;
-}
