--- apps/skeleton-app/src/_gqlTypes/ME.ts
@@ -1,44 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-/* tslint:disable */
-/* eslint-disable */
-// @generated
-// This file was automatically generated and should not be edited.
-
-import {LibraryBehavior} from './globalTypes';
-
-// ====================================================
-// GraphQL query operation: ME
-// ====================================================
-
-export interface ME_me_whoAmI_library_gqlNames {
-    query: string;
-    type: string;
-}
-
-export interface ME_me_whoAmI_library {
-    id: string;
-    behavior: LibraryBehavior;
-    label: SystemTranslation | null;
-    gqlNames: ME_me_whoAmI_library_gqlNames;
-}
-
-export interface ME_me_whoAmI {
-    id: string;
-    label: string | null;
-    subLabel: string | null;
-    color: string | null;
-    library: ME_me_whoAmI_library;
-    preview: Preview | null;
-}
-
-export interface ME_me {
-    login: string | null;
-    id: string;
-    whoAmI: ME_me_whoAmI;
-}
-
-export interface ME {
-    me: ME_me | null;
-}
