--- apps/skeleton-app/src/_gqlTypes/RecordIdentity.ts
@@ -1,34 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-/* tslint:disable */
-/* eslint-disable */
-// @generated
-// This file was automatically generated and should not be edited.
-
-import {LibraryBehavior} from './globalTypes';
-
-// ====================================================
-// GraphQL fragment: RecordIdentity
-// ====================================================
-
-
-export interface RecordIdentity_whoAmI_library {
-    id: string;
-    behavior: LibraryBehavior;
-    label: SystemTranslation | null;
-}
-
-export interface RecordIdentity_whoAmI {
-    id: string;
-    label: string | null;
-    subLabel: string | null;
-    color: string | null;
-    library: RecordIdentity_whoAmI_library;
-    preview: Preview | null;
-}
-
-export interface RecordIdentity {
-    id: string;
-    whoAmI: RecordIdentity_whoAmI;
-}
