--- apps/skeleton-app/src/graphQL/queries/applications/applicationDetailsFragment.ts
@@ -1,27 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {gql} from '@apollo/client';
-import recordIdentityFragment from '../records/recordIdentityFragment';
-
-export const applicationDetailsFragment = gql`
-    ${recordIdentityFragment}
-    fragment ApplicationDetails on Application {
-        id
-        label
-        type
-        description
-        endpoint
-        url
-        color
-        icon {
-            ...RecordIdentity
-        }
-        module
-        permissions {
-            access_application
-            admin_application
-        }
-        settings
-    }
-`;
