--- apps/skeleton-app/src/graphQL/queries/applications/getApplicationByEndpointQuery.ts
@@ -1,16 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {gql} from '@apollo/client';
-import {applicationDetailsFragment} from './applicationDetailsFragment';
-
-export const getApplicationByEndpointQuery = gql`
-    ${applicationDetailsFragment}
-    query GET_APPLICATION_BY_ENDPOINT($endpoint: String!) {
-        applications(filters: {endpoint: $endpoint}) {
-            list {
-                ...ApplicationDetails
-            }
-        }
-    }
-`;
