--- apps/skeleton-app/src/graphQL/queries/globalSettings/getGlobalSettingsQuery.ts
@@ -1,18 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import gql from 'graphql-tag';
-import recordIdentityFragment from '../records/recordIdentityFragment';
-
-export const getGlobalSettingsQuery = gql`
-    ${recordIdentityFragment}
-    query GET_GLOBAL_SETTINGS {
-        globalSettings {
-            name
-            icon {
-                id
-                ...RecordIdentity
-            }
-        }
-    }
-`;
