--- apps/skeleton-app/src/graphQL/queries/records/recordIdentityFragment.ts
@@ -1,24 +1,0 @@
-// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {gql} from 'graphql-tag';
-
-const recordIdentityFragment = gql`
-    fragment RecordIdentity on Record {
-        id
-        whoAmI {
-            id
-            label
-            subLabel
-            color
-            library {
-                id
-                behavior
-                label
-            }
-            preview
-        }
-    }
-`;
-
-export default recordIdentityFragment;
