+++ apps/skeleton-app/src/modules/Initializer/languages.graphql
@@ -1,0 +1,3 @@
+query getLanguages {
+    langs
+}
